6 acceptance tests

+++ app/controllers/questions_controller.rb
@@ -1,10 +1,15 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: [:index, :show]
+
   expose :questions, ->{ Question.all }
   expose :question
+  expose :answer, ->{ Answer.new }
 
   def create
+    question.author = current_user
+
     if question.save
-      redirect_to question
+      redirect_to question, notice: 'Your question successfully created.'
     else
       render :new
     end
@@ -19,7 +24,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
+    if current_user.author_of?(question)
+      question.destroy
+      flash[:notice] = 'Your question successfully deleted.'
+    else
+      flash[:notice] = "You can not delete another user's question"
+    end
     redirect_to questions_path
   end
